add: Executive section with util function to extract team data from MD file
