+++ src/lib/utils.ts
@@ -1,9 +1,9 @@
 import { type ClassValue, clsx } from "clsx";
-import { twMerge } from "tailwind-merge";
 import fs from "fs";
-import path from "path";
 import matter from "gray-matter";
-import { TeamMemberProps, TEvent } from "./types";
+import path from "path";
+import { twMerge } from "tailwind-merge";
+import { TeamsProps, TEvent } from "./types";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -61,7 +61,7 @@ export function getSortedPostsData() {
 }
 
 // Function to read and parse the team markdown file
-export const getTeamDataByDivision = (): TeamMemberProps => {
+export const getTeamDataByDivision = (): TeamsProps => {
   const executiveTeamDataDir = path.join(
     process.cwd(),
     "src/data/team/ExecutiveTeam.data.md"
@@ -69,7 +69,7 @@ export const getTeamDataByDivision = (): TeamMemberProps => {
   const fileContent = fs.readFileSync(executiveTeamDataDir, "utf8");
 
   // Use gray-matter to parse the front matter
-  const { data } = matter(fileContent) as unknown as { data: TeamMemberProps };
+  const { data } = matter(fileContent) as unknown as { data: TeamsProps };
 
   // Return the parsed team data by division
   return data;
